refactor: split oxlint-config.ts (722 LOC) into src/core/runners/oxlint/
Same shape as the scan.ts split: the 722-line oxlint-config.ts mixed
five distinct concerns. Carved into six focused files under
src/core/runners/oxlint/, with oxlint-config.ts kept as a thin
re-export barrel so existing importers (run-oxlint.ts, eslint-plugin.ts,
merge-and-filter-diagnostics.ts, tests) keep working unchanged.

New layout:

  src/core/runners/oxlint/
    types.ts             11 LOC  RuleSeverity + OxlintConfigOptions
    rule-maps.ts        305 LOC  per-framework severity maps (NEXTJS_RULES,
                                 REACT_NATIVE_RULES, TANSTACK_START_RULES,
                                 TANSTACK_QUERY_RULES, GLOBAL_REACT_DOCTOR_RULES,
                                 REACT_COMPILER_RULES, YOU_MIGHT_NOT_NEED_EFFECT_RULES,
                                 BUILTIN_REACT_RULES, BUILTIN_A11Y_RULES,
                                 ALL_REACT_DOCTOR_RULE_KEYS, FRAMEWORK_SPECIFIC_RULE_KEYS)
    rule-metadata.ts    166 LOC  RULE_METADATA + tag constants + RuleMetadataEntry
    plugin-resolution.ts 110 LOC  resolveReactHooksJsPlugin / resolve-You-Might-...
                                  / readPluginRuleNames / filterRulesToAvailable
    capabilities.ts      56 LOC  buildCapabilities + shouldEnableRule
    config.ts            98 LOC  createOxlintConfig (the composer that pulls
                                 everything together)

  src/core/runners/oxlint-config.ts  12 LOC  re-export barrel

Compared to before: 722-line monolithic file -> 758 LOC total across 7
focused files, each under 305 LOC.

Internal-only types/helpers that were exported unnecessarily are now
unexported in their new homes (e.g. RuleMetadataEntry stays internal to
rule-metadata.ts; buildCapabilities/shouldEnableRule are internal to
the capabilities module).

Verification:
  - typecheck/lint/format clean
  - 734/734 tests pass
  - No public API change (barrel exposes the same surface)
  - run-oxlint.ts, eslint-plugin.ts, merge-and-filter-diagnostics.ts,
    and tests all import from the unchanged oxlint-config.js barrel
