+++ packages/react-doctor/src/core/runners/oxlint/types.ts
@@ -1,0 +1,11 @@
+import type { ProjectInfo } from "../../../types.js";
+
+export type RuleSeverity = "error" | "warn" | "off";
+
+export interface OxlintConfigOptions {
+  pluginPath: string;
+  project: ProjectInfo;
+  customRulesOnly?: boolean;
+  extendsPaths?: string[];
+  ignoredTags?: ReadonlySet<string>;
+}
